refactor(intake): nest package rule subsystem
Why:
- move the intake package-rule cluster into the next bounded subpackage
- keep parity proof paths aligned with the new intake feature layout

What:
- move package-rule modules under intake/packages and keep the intake public exports stable
- mirror the package-rule tests under tests/unit/application/services/intake/packages
- update parity docs and standards references to the nested package paths

Checks:
- uv run pytest --no-cov tests/unit/application/services/intake/packages tests/unit/application/services/intake/test_plan_package_review_workflow.py
- uv run pre-commit run markdownlint --all-files
- uv run python -m tools.run_quality_gates --full-tests
