+++ docs/MASTER_PARITY_LEDGER.md
@@ -35,14 +35,14 @@ or file locations are not revived to satisfy parity.
 | ---- | ----: | ---- | ---- | ---- |
 | `tests/adapters/test_adapter_protocol.py` | 1 | `tests/contract/test_adapter_registry.py` | `superseded-direct` | Replaced by package-style registry and contract discovery tests |
 | `tests/e2e/test_scripts.py` | 13 | `tests/e2e/test_cli.py` | `superseded-direct` | CLI workflows remain active under the typed command surface |
-| `tests/pipeline/test_intake_sort.py` | 17 | `tests/unit/application/services/intake/test_apply_workflow.py`; `tests/unit/application/services/intake/test_plan_package_review_workflow.py`; `tests/unit/application/services/intake/test_plan_routing_workflow.py`; `tests/unit/application/services/intake/test_package_cycles.py`; `tests/unit/application/services/intake/test_package_duplicates.py`; `tests/unit/application/services/intake/test_package_isolation.py`; `tests/unit/application/services/intake/test_package_merges.py`; `tests/unit/application/services/intake/test_package_scope_conflicts.py`; `tests/unit/application/services/intake/test_package_scope_precedence.py` | `ported-direct` | Legacy sort and grouping rules now live in typed intake planning and package-resolution seams under the dedicated intake package |
+| `tests/pipeline/test_intake_sort.py` | 17 | `tests/unit/application/services/intake/test_apply_workflow.py`; `tests/unit/application/services/intake/test_plan_package_review_workflow.py`; `tests/unit/application/services/intake/test_plan_routing_workflow.py`; `tests/unit/application/services/intake/packages/test_cycles.py`; `tests/unit/application/services/intake/packages/test_duplicates.py`; `tests/unit/application/services/intake/packages/test_isolation.py`; `tests/unit/application/services/intake/packages/test_merges.py`; `tests/unit/application/services/intake/packages/test_scope_conflicts.py`; `tests/unit/application/services/intake/packages/test_scope_precedence.py` | `ported-direct` | Legacy sort and grouping rules now live in typed intake planning and package-resolution seams under the dedicated intake package |
 | `tests/unit/test_baseline_check.py` | 19 | `tests/unit/application/services/test_baseline_service.py`; `tests/unit/application/services/test_batch_staging_workflow.py`; `tests/contract/test_runtime_artifacts.py` | `ported-direct` | Baseline validation, artifact writing, exchange-only assets, and balance-only source handling are directly covered |
 | `tests/unit/test_binance_unwrap.py` | 5 | `tests/unit/adapters/test_binance_adapter.py`; `tests/contract/test_source_adapter_packs.py`; `tests/contract/test_source_pack_profiles.py` | `ported-direct` | Includes filename offset parsing, sentinel no-data handling, spot/deposit/withdraw/staking mapping, and pack-level normalization proof |
 | `tests/unit/test_coinbase_common.py` | 4 | `tests/unit/adapters/test_coinbase_adapter.py`; `tests/unit/application/services/test_pdf_extract_service.py`; `tests/unit/application/services/test_supporting_services.py` | `ported-direct` | Retail parsing, renamed retail exports, transaction mapping, and statement balance extraction remain covered |
 | `tests/unit/test_inspection.py` | 17 | `tests/unit/application/services/intake/test_archive_scan_workflow.py`; `tests/unit/application/services/intake/test_archive_scan_limits.py`; `tests/unit/application/services/intake/test_file_facts.py`; `tests/unit/application/services/test_profile_inventory_details.py`; `tests/unit/application/services/test_manifest_service.py` | `ported-direct` | Archive inspection, delimiter/date inference, artifact classification, and manifest/profile inventory behavior are direct seams |
 | `tests/unit/test_normalization_common.py` | 7 | `tests/unit/test_domain_value_objects.py`; `tests/unit/application/services/test_verification_summary.py`; `tests/unit/application/services/test_workflow_services.py` | `superseded-direct` | Decimal and canonical rendering helpers moved into typed domain and verification seams |
 | `tests/unit/test_overlap_check.py` | 9 | `tests/unit/application/services/test_overlap_service.py`; `tests/unit/application/services/test_workflow_services.py` | `ported-direct` | Candidate overlap parsing, summaries, flagged rows, and artifact writing are directly covered |
-| `tests/unit/test_package_resolution.py` | 23 | `tests/unit/application/services/intake/test_package_cycles.py`; `tests/unit/application/services/intake/test_package_duplicates.py`; `tests/unit/application/services/intake/test_package_helpers.py`; `tests/unit/application/services/intake/test_package_isolation.py`; `tests/unit/application/services/intake/test_package_merges.py`; `tests/unit/application/services/intake/test_package_scope_conflicts.py`; `tests/unit/application/services/intake/test_package_scope_precedence.py`; `tests/unit/application/services/intake/test_plan_package_review_workflow.py` | `ported-direct` | Same-cycle merges, mixed-cycle review, scope conflicts, subset elimination, and overlap decisions are directly covered inside the intake feature package |
+| `tests/unit/test_package_resolution.py` | 23 | `tests/unit/application/services/intake/packages/test_cycles.py`; `tests/unit/application/services/intake/packages/test_duplicates.py`; `tests/unit/application/services/intake/packages/test_helpers.py`; `tests/unit/application/services/intake/packages/test_isolation.py`; `tests/unit/application/services/intake/packages/test_merges.py`; `tests/unit/application/services/intake/packages/test_scope_conflicts.py`; `tests/unit/application/services/intake/packages/test_scope_precedence.py`; `tests/unit/application/services/intake/test_plan_package_review_workflow.py` | `ported-direct` | Same-cycle merges, mixed-cycle review, scope conflicts, subset elimination, and overlap decisions are directly covered inside the intake feature package |
 | `tests/unit/test_pdf_balance_extract.py` | 3 | `tests/unit/application/services/test_pdf_extract_service.py`; `tests/unit/application/services/test_supporting_services.py` | `ported-direct` | Statement detection and supported PDF balance extraction remain present |
 | `tests/unit/test_pipeline_common.py` | 12 | `tests/unit/application/services/test_export_files.py`; `tests/unit/test_serialization_io.py`; `tests/unit/application/services/test_verification_summary.py`; `tests/unit/application/services/test_rounds_service.py` | `superseded-direct` | CSV/JSON IO, required export lookup, default verification export ordering, and formatting helpers are now split into typed seams |
 | `tests/unit/test_reconcile_source.py` | 3 | `tests/unit/application/services/test_supporting_services.py`; `tests/unit/application/services/test_verification_compare_workflow.py` | `ported-direct` | Reconciliation diffs and workflow outputs remain covered |
