+++ docs/MASTER_TEST_PARITY_MATRIX.md
@@ -15,14 +15,14 @@ Statuses:
 | ---- | ---- | ---- |
 | `tests/adapters/test_adapter_protocol.py` | `covered` | `tests/contract/test_adapter_registry.py` |
 | `tests/e2e/test_scripts.py` | `ported` | `tests/e2e/test_cli.py`, `tests/unit/adapters/test_adapter_tooling.py`, `tests/unit/application/services/test_workflow_services.py` |
-| `tests/pipeline/test_intake_sort.py` | `ported` | `tests/unit/application/services/test_intake_service.py`, `tests/unit/test_intake_packages.py`, `tests/unit/application/services/test_intake_routing.py`, `tests/unit/application/services/test_intake_file_facts.py` |
+| `tests/pipeline/test_intake_sort.py` | `ported` | `tests/unit/application/services/intake/test_apply_workflow.py`, `tests/unit/application/services/intake/test_plan_package_review_workflow.py`, `tests/unit/application/services/intake/test_plan_routing_workflow.py`, `tests/unit/application/services/intake/packages/test_cycles.py`, `tests/unit/application/services/intake/packages/test_duplicates.py`, `tests/unit/application/services/intake/packages/test_isolation.py`, `tests/unit/application/services/intake/packages/test_merges.py`, `tests/unit/application/services/intake/packages/test_scope_conflicts.py`, `tests/unit/application/services/intake/packages/test_scope_precedence.py` |
 | `tests/unit/test_baseline_check.py` | `ported` | `tests/unit/application/services/test_baseline_service.py`, `tests/e2e/test_cli.py` |
 | `tests/unit/test_binance_unwrap.py` | `superseded-with-proof` | `tests/unit/application/services/test_archive_scan.py`, `tests/unit/application/services/test_intake_service.py`, `tests/unit/adapters/test_binance_adapter.py` |
 | `tests/unit/test_coinbase_common.py` | `ported` | `tests/unit/adapters/test_coinbase_adapter.py`, `tests/unit/application/services/test_supporting_services.py` |
 | `tests/unit/test_inspection.py` | `ported` | `tests/unit/application/services/test_archive_scan.py`, `tests/unit/application/services/test_intake_file_facts.py`, `tests/unit/application/services/test_intake_routing.py` |
 | `tests/unit/test_normalization_common.py` | `superseded-with-proof` | `tests/unit/adapters/test_shared_support.py`, `tests/unit/test_hardening.py` |
 | `tests/unit/test_overlap_check.py` | `ported` | `tests/unit/application/services/test_workflow_services.py`, `tests/e2e/test_cli.py` |
-| `tests/unit/test_package_resolution.py` | `covered` | `tests/unit/test_intake_packages.py` |
+| `tests/unit/test_package_resolution.py` | `covered` | `tests/unit/application/services/intake/packages/test_cycles.py`, `tests/unit/application/services/intake/packages/test_duplicates.py`, `tests/unit/application/services/intake/packages/test_helpers.py`, `tests/unit/application/services/intake/packages/test_isolation.py`, `tests/unit/application/services/intake/packages/test_merges.py`, `tests/unit/application/services/intake/packages/test_scope_conflicts.py`, `tests/unit/application/services/intake/packages/test_scope_precedence.py`, `tests/unit/application/services/intake/test_plan_package_review_workflow.py` |
 | `tests/unit/test_pdf_balance_extract.py` | `ported` | `tests/unit/application/services/test_supporting_services.py` |
 | `tests/unit/test_pipeline_common.py` | `ported` | `tests/unit/application/services/test_archive_scan.py`, `tests/unit/application/services/test_intake_file_facts.py`, `tests/unit/application/services/test_intake_routing.py` |
 | `tests/unit/test_reconcile_source.py` | `ported` | `tests/unit/application/services/test_supporting_services.py` |
