+++ src/crypto_reconciliation/application/services/intake/packages/resolution.py
@@ -6,22 +6,11 @@ from collections import Counter, defaultdict
 from dataclasses import replace
 from datetime import datetime
 
-from .package_duplicates import apply_duplicate_decisions
-from .package_markers import (
-    logical_key,
-    material_indexes,
-    package_sort_key,
-    row_marker,
-)
-from .package_merges import apply_merge_decisions
-from .package_models import (
-    BundlePackage,
-    PackageGroupKey,
-    PackageRuleSummary,
-    PlannedPackageItem,
-    package_key,
-)
-from .package_reviews import apply_default_decisions, apply_overlap_review_decisions
+from .duplicates import apply_duplicate_decisions
+from .markers import logical_key, material_indexes, package_sort_key, row_marker
+from .merges import apply_merge_decisions
+from .models import BundlePackage, PackageGroupKey, PackageRuleSummary, PlannedPackageItem, package_key
+from .reviews import apply_default_decisions, apply_overlap_review_decisions
 
 
 def apply_package_rules(
